switched graph to use echarts-for-react
