+++ packages/dina-ui/package.json
@@ -19,6 +19,7 @@
     "coordinate-parser": "^1.0.7",
     "dexie": "^3.2.4",
     "dexie-react-hooks": "^1.1.6",
+    "echarts-for-react": "^3.0.6",
     "esri-loader": "^3.3.0",
     "esri-loader-hooks": "^0.4.0",
     "moment": "^2.29.4",
